requirements.txt
Flask==0.12.2
line-bot-sdk==1.8.0
beautifulsoup4==4.8.1
requests==2.23.0